requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
PyJWT==2.8.0
pytz==2024.1
referencing==0.35.1
rpds-py==0.19.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
